+++ src/led/headers/led.h
@@ -23,7 +23,7 @@ class LEDS
         };
         static const uint16_t PWM_RANGE = 4096;
         LEDS(uint8_t Pin, uint8_t Mode);
-        void writePwm(uint16_t DutyCycle);
+        void writePwm(int DutyCycle);
         void writeDigital(bool Status);
         bool readDigital();
         bool getDigitalOutStatus();
